scraper/requirements.txt
playwright
pandas
pyarrow
requests
Pillow